feat: allow login to non-db contracts
Before this commit, each time updateAccount was triggered the VM checked if the selector was connected - i.e. had an access key - to a contract different from SocialDB, and proceed to logout the user if that was the case.

This commit removes this constraint, allowing users to connect to both SocialDB and other contracts.